make heartbeat interval (int -> float) for split seconds interval

+++ ants/base_config.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass
 
 @dataclass
 class BaseConfig:
-    heartbeat_interval: int  # In seconds
+    heartbeat_interval: float  # In seconds
